+++ hubble.backend.business.services/src/main/java/hubble/backend/business/services/interfaces/operations/kpis/EventKpiOperations.java
@@ -12,7 +12,7 @@ public interface EventKpiOperations extends
 
     public EventsKpi calculateLastMonthKeyPerformanceIndicatorByApplication(String applicationId);
 
-    public long calculateLastDayKPI(String applicationId);
+    public double calculateLastDayKPI(String applicationId);
 
-    public long calculatePastDayKPI(String applicationId);
+    public double calculatePastDayKPI(String applicationId);
 }
