+++ hubble.backend.business.services/src/main/java/hubble/backend/business/services/interfaces/services/WorkItemService.java
@@ -13,9 +13,9 @@ public interface WorkItemService extends ServiceBase<WorkItem>,
 
     public WorkItemQuantity calculateWorkItemQuantityLastWeek(String applicationId);
 
-    public long calculateLastDayDeflectionDaysKpi(String applicationId);
+    public double calculateLastDayDeflectionDaysKpi(String applicationId);
 
-    public long calculatePastDayDeflectionDaysKpi(String applicationId);
+    public double calculatePastDayDeflectionDaysKpi(String applicationId);
 
     List<Integer> getDistValuesLastDay(String id);
 }
